requirements.txt
Appium-Python-Client==2.1.0
pytest==6.2.5
